# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TEST OpenApi  Документація",
        "description": "L5 Swagger OpenApi опис",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/currencies": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "summary": "All currencies",
                "description": "Return json list currencies",
                "operationId": "d19d1d4d322c499561dbceb10ce63a3a",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/currencies/store": {
            "post": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Creates a new currency",
                "description": "returns status",
                "operationId": "cc3029b940d3e1f523045de95f3a39b0",
                "requestBody": {
                    "description": "Data packet for Test",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Доллар США"
                                    },
                                    "short_name": {
                                        "type": "string",
                                        "example": "доллар"
                                    },
                                    "name_iso": {
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "code_iso": {
                                        "type": "string",
                                        "example": "840"
                                    },
                                    "symbol_iso": {
                                        "type": "string",
                                        "example": "$"
                                    },
                                    "name1": {
                                        "type": "string",
                                        "example": "Доллар"
                                    },
                                    "name2": {
                                        "type": "string",
                                        "example": "Доллара"
                                    },
                                    "name3": {
                                        "type": "string",
                                        "example": "Долларов"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/currencies/show/{id}": {
            "get": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Single currency data",
                "description": "Returns json",
                "operationId": "279221e9fcfa9a9423972bf1bee24c85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/currencies/update/{id}": {
            "put": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Зміна даних клієнта",
                "description": "Повертає статус",
                "operationId": "a7ccfecdca7ce1d80dd7cb1724a36274",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id акаунта",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data packet for Test",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Доллар США"
                                    },
                                    "short_name": {
                                        "type": "string",
                                        "example": "доллар"
                                    },
                                    "name_iso": {
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "code_iso": {
                                        "type": "string",
                                        "example": "840"
                                    },
                                    "symbol_iso": {
                                        "type": "string",
                                        "example": "$"
                                    },
                                    "name1": {
                                        "type": "string",
                                        "example": "Доллар"
                                    },
                                    "name2": {
                                        "type": "string",
                                        "example": "Доллара"
                                    },
                                    "name3": {
                                        "type": "string",
                                        "example": "Долларов"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/currencies/destroy/{id}": {
            "delete": {
                "tags": [
                    "Currencies"
                ],
                "summary": "Видалити клієнта",
                "description": "Повертає статус",
                "operationId": "b023320675bd9195d9d49cd137a6c936",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}